+++ package.json
@@ -11,6 +11,7 @@
     "generate": "nuxt generate"
   },
   "dependencies": {
+    "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "googleapis": "^49.0.0",
     "nuxt": "^2.0.0"
